split multimodal tests
