+++ .github/workflow_scripts/test_multimodal_predictor.sh
