+++ multimodal/tests/unittests/others/test_auto_model.py
