+++ multimodal/tests/unittests/others/test_backward_compatibility.py
@@ -4,9 +4,9 @@ import tempfile
 
 import numpy.testing as npt
 import pytest
-from test_predictor import verify_predictor_save_load
-from unittest_datasets import AmazonReviewSentimentCrossLingualDataset
-from utils import get_home_dir, protected_zip_extraction
+from ..predictor.test_predictor import verify_predictor_save_load
+from .unittest_datasets import AmazonReviewSentimentCrossLingualDataset
+from .utils import get_home_dir, protected_zip_extraction
 
 from autogluon.multimodal import MultiModalPredictor
 from autogluon.multimodal.utils import download
