+++ multimodal/tests/unittests/others/test_data_augmentation.py
@@ -4,8 +4,8 @@ import pickle
 import shutil
 import tempfile
 
-from test_predictor import verify_predictor_save_load
-from unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
+from ..predictor.test_predictor import verify_predictor_save_load
+from .unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
 
 from autogluon.multimodal import MultiModalPredictor
 from autogluon.multimodal.constants import (
