+++ multimodal/tests/unittests/others/test_data_process_image.py
@@ -1,6 +1,6 @@
 import PIL
 import pytest
-from unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
+from .unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
 
 from autogluon.multimodal.data.process_image import ImageProcessor
 from autogluon.multimodal.models import TimmAutoModelForImagePrediction
