+++ multimodal/tests/unittests/others/test_distiller.py
@@ -1,9 +1,9 @@
 import os
 import shutil
 
-from test_predictor import verify_predictor_save_load
-from unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
-from utils import get_home_dir
+from ..predictor.test_predictor import verify_predictor_save_load
+from .unittest_datasets import AEDataset, HatefulMeMesDataset, PetFinderDataset
+from .utils import get_home_dir
 
 from autogluon.multimodal import MultiModalPredictor
