+++ multimodal/tests/unittests/others/test_object_detection.py
