+++ multimodal/tests/unittests/others/test_registry.py
