+++ multimodal/tests/unittests/others/test_save_path.py
@@ -2,7 +2,7 @@ import os
 import shutil
 
 import pytest
-from unittest_datasets import PetFinderDataset
+from .unittest_datasets import PetFinderDataset
 
 from autogluon.multimodal import MultiModalPredictor
