+++ multimodal/tests/unittests/others/test_text_detection.py
