+++ multimodal/tests/unittests/others/test_text_recognition.py
