+++ multimodal/tests/unittests/others/test_zero_shot.py
